Refactor Medicine.java. Change Setter to Builder template.
